[DEV-14329] despark II
