+++ usaspending_api/common/data_classes.py
@@ -1,7 +1,6 @@
 from dataclasses import dataclass
-from typing import Callable, Optional
+from typing import Optional
 
-from pyspark.sql.types import StructType
 from typing_extensions import Literal
 
 
@@ -50,36 +49,3 @@ class TransactionColumn:
     #   to be applied on this input. For example, a valid scalar_transformation string is
     #   "CASE {input} WHEN 'UNITED STATES' THEN 'USA' ELSE {input} END"
     scalar_transformation: str = None
-
-
-@dataclass(slots=True)
-class TableSpec:
-    """
-    Delta table metadata class
-    """
-
-    destination_database: str
-    delta_table_create_sql: str | StructType
-
-    destination_table_name: str | None = None
-    source_table: str | None = None
-    source_database: str | None = None
-    model: str | None = None
-    partition_column: str | None = None
-    partition_column_type: str | None = None
-    is_partition_column_unique: bool = False
-    delta_table_create_options: dict | None = None
-    delta_table_create_partitions: list[str] | None = None
-    source_schema: list | dict | None = None
-    custom_schema: str | None = None
-    column_names: list[str] | None = None
-    is_from_broker: bool = False
-    source_query: str | list[str] | Callable | None = None
-    source_query_incremental: str | Callable | None = None
-    user_defined_functions: list[dict] | None = None
-    archive_data_field: str = "update_date"
-    postgres_seq_name: str | None = None
-    postgres_partition_spec: dict | None = None
-    tsvectors: list[str] | dict | None = None
-    swap_table: str | None = None
-    swap_schema: str | None = None
