+++ usaspending_api/etl/management/commands/archive_table_in_delta.py
@@ -5,7 +5,6 @@ from datetime import datetime, timedelta
 import psycopg
 from django.core.management.base import BaseCommand
 
-from usaspending_api.common.data_classes import TableSpec
 from usaspending_api.common.etl.spark import load_delta_table
 from usaspending_api.common.helpers.spark_helpers import (
     configure_spark_session,
@@ -14,6 +13,7 @@ from usaspending_api.common.helpers.spark_helpers import (
     get_usas_jdbc_url,
 )
 from usaspending_api.common.helpers.sql_helpers import get_database_dsn_string
+from usaspending_api.common.spark_data_classes import TableSpec
 from usaspending_api.download.delta_models.download_job import (
     download_job_create_sql_string,
 )
